KEYS: Add a keyctl function to allow a key to be altered in type-dependent way
Add an operation to permit a key to be altered in a type-dependent way.  This
is given a text string command that indicates the alteration to be made.  The
operation may also be given a blob of data to supplement the command.

	long keyctl_alter(key_serial_t keyid,
			  const char *command,
			  const void *data, size_t data_size);

The caller must have WRITE permission on a key to be able to perform this
operation.  The type is not required to implement this, but if it does, it must
perform its own locking against other 'writes' using the key semaphore.

The total command size can be up to 4KB in size including the NUL terminator.
The command must be prefaced with the intended key type name and a space.  This
does not, however, preclude types from implementing alterations intended for
other types.

The data buffer is optional and can be up to 1MB in size.

This can be invoked from the keyctl command in userspace.  One example would be
to use this to change the master key used by an encrypted key:

	keyctl alter 1234 "encrypted change-master-key" 6789

The alteration operation is different to the update function that simply
replaces the entire key payload.

Signed-off-by: David Howells <dhowells@redhat.com>