sign-file: Document dependency on OpenSSL devel libraries
The revised sign-file program is no longer a script that wraps the openssl
program, but now rather a program that makes use of the OpenSSL's crypto
library.  This means that to build the sign-file program, the kernel build
process now has a dependency on openssl-devel in addition to openssl.

Document this in Kconfig and in module-signing.txt.

Signed-off-by: David Howells <dhowells@redhat.com>