rxrpc: Don't permit use of connect() op and simplify sendmsg() op
Simplify the RxRPC user interface and remove the use of connect() to direct
client calls.  It is redundant given that sendmsg() can be given the target
address and calls to multiple targets are permitted from a client socket
and also from a service socket.

Simplify sendmsg() also.  If we can't find a call immediately, we create
one, as now, but if the call then exists when we try and add it, we give an
error rather than using the call we found at the second attempt.  We should
never see this situation unless two threads are racing, trying to create a
call with the same ID - which would be an error.

It also isn't required to provide sendmsg() with an address - provided the
control message data holds a user ID that maps to a currently active call.

Signed-off-by: David Howells <dhowells@redhat.com>